requirements.txt
altair==5.1.2
annotated-types==0.6.0
appdirs==1.4.4
astor==0.8.1
asttokens==2.4.1
attrs==23.1.0
azure-ai-textanalytics==5.2.0
azure-common==1.1.28
azure-core==1.29.5
beautifulsoup4==4.12.2
blinker==1.7.0
blis==0.7.11
cachetools==4.2.4
catalogue==2.0.10
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
comm==0.1.4
confection==0.1.3
contourpy==1.2.0
cycler==0.12.1
cymem==2.0.8
debugpy==1.8.0
decorator==5.1.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl#sha256=6215d71a3212690e9aec49408a27e3fe6ad7cd6c715476e93d70dc784041e93e
es-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/es_core_news_sm-3.7.0/es_core_news_sm-3.7.0-py3-none-any.whl#sha256=61e6e5530941f5880166855f09f60d7e6ba79ec1e8e45f96244bdb1eb169eb1d
executing==2.0.1
filelock==3.13.1
fonttools==4.44.0
fsspec==2023.10.0
gitdb==4.0.11
GitPython==3.1.40
google-api-core==1.34.0
google-api-python-client==2.106.0
google-auth==1.35.0
google-auth-httplib2==0.1.1
google-cloud-core==1.7.3
google-cloud-translate==2.0.1
googleapis-common-protos==1.61.0
grpcio==1.59.2
grpcio-status==1.48.2
h11==0.14.0
httplib2==0.22.0
huggingface-hub==0.19.4
idna==3.4
importlib-metadata==6.8.0
ipykernel==6.26.0
ipython==8.17.2
isodate==0.6.1
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter_client==8.5.0
jupyter_core==5.5.0
kiwisolver==1.4.5
langcodes==3.3.0
langid==1.1.6
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
msrest==0.7.1
murmurhash==1.0.10
nest-asyncio==1.5.8
networkx==3.2.1
numpy==1.26.1
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.3
pandasgui==0.2.14
parso==0.8.3
Pillow==10.1.0
platformdirs==3.11.0
plotly==5.18.0
preshed==3.0.9
prompt-toolkit==3.0.40
protobuf==3.20.3
psutil==5.9.6
pure-eval==0.2.2
pyarrow==14.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pydeck==0.8.1b0
Pygments==2.16.1
pynput==1.7.6
pyparsing==3.1.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
PyQtWebEngine==5.15.6
PyQtWebEngine-Qt5==5.15.2
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.1
qtstylish==0.1.5
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.6.0
rpds-py==0.12.0
rsa==4.9
sacremoses==0.1.1
safetensors==0.4.0
selenium==4.15.2
sentencepiece==0.1.99
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
streamlit==1.28.1
sympy==1.12
tenacity==8.2.3
thinc==8.2.1
tokenizers==0.15.0
toml==0.10.2
toolz==0.12.0
torch==2.1.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
transformers==4.35.2
trio==0.23.1
trio-websocket==0.11.1
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.0.7
validators==0.22.0
wasabi==1.1.2
watchdog==3.0.0
wcwidth==0.2.9
weasel==0.3.3
wordcloud==1.9.2
wsproto==1.2.0
zipp==3.17.0
